fixed checkbox functionality
